# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bidly API Documentation",
        "description": "행사 기획 대행사와 용역사를 연결하는 입찰 플랫폼 API",
        "contact": {
            "name": "Bidly Support Team",
            "email": "support@bidly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/contracts/{contract}/propose-meeting-dates": {
            "post": {
                "tags": [
                    "Contract Meetings"
                ],
                "summary": "미팅 일정 제안",
                "description": "대행사가 용역사에게 미팅 일정을 제안합니다",
                "operationId": "1686b1b41cd7a53a08450dd101e2a392",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "description": "계약 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "proposed_dates"
                                ],
                                "properties": {
                                    "proposed_dates": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "datetime": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "회사 회의실"
                                                },
                                                "note": {
                                                    "type": "string",
                                                    "example": "계약 조건 협의"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "미팅 일정 제안 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "미팅 일정이 성공적으로 제안되었습니다"
                                        },
                                        "proposed_dates_count": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "계약을 찾을 수 없습니다"
                    },
                    "422": {
                        "description": "유효하지 않은 입력 데이터"
                    }
                }
            }
        },
        "/api/contracts/{contract}/select-meeting-date": {
            "patch": {
                "tags": [
                    "Contract Meetings"
                ],
                "summary": "미팅 일정 선택",
                "description": "용역사가 제안된 미팅 일정 중 하나를 선택합니다",
                "operationId": "f9a8e7a2fe9bdddb8959401176997021",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "description": "계약 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "selected_datetime"
                                ],
                                "properties": {
                                    "selected_datetime": {
                                        "type": "string",
                                        "format": "datetime"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "회사 회의실"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "계약 조건 협의"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "미팅 일정 선택 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "미팅 일정이 성공적으로 확정되었습니다"
                                        },
                                        "schedule_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "계약을 찾을 수 없습니다"
                    },
                    "422": {
                        "description": "유효하지 않은 입력 데이터"
                    }
                }
            }
        },
        "/api/contracts/{contract}/meeting-status": {
            "get": {
                "tags": [
                    "Contract Meetings"
                ],
                "summary": "미팅 상태 조회",
                "description": "계약의 미팅 진행 상태를 조회합니다",
                "operationId": "36caaf869d9728f6959a57faa212a023",
                "parameters": [
                    {
                        "name": "contract",
                        "in": "path",
                        "description": "계약 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "미팅 상태 조회 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "contract_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "meeting_status": {
                                            "type": "string",
                                            "example": "date_selected"
                                        },
                                        "proposed_meeting_dates": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "selected_meeting_date": {
                                            "type": "string",
                                            "format": "datetime"
                                        },
                                        "schedule": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "계약을 찾을 수 없습니다"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}